beginning a rebase

+++ backend-api/src/main/java/com/team2/spartanslist/order/OrderController.java
@@ -5,7 +5,9 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -16,6 +18,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import com.team2.spartanslist.Global;
 import com.team2.spartanslist.cart.CartService;
 import com.team2.spartanslist.offer.OfferService;
+import com.team2.spartanslist.review.Review;
+import com.team2.spartanslist.seller.Seller;
+import com.team2.spartanslist.seller.SellerService;
 import com.team2.spartanslist.offer.Offer;
 import com.team2.spartanslist.shopper.Shopper;
 import com.team2.spartanslist.shopper.ShopperService;
@@ -35,6 +40,8 @@ public class OrderController {
     ShopperService shopperService;
     @Autowired
     CartService cartService;
+    @Autowired
+    SellerService sellerService;
 
     @PostMapping("/{offerID}")
     public String createOrder(@PathVariable Long offerID, Authentication auth){
@@ -47,7 +54,7 @@ public class OrderController {
         Offer offer = offerService.getOfferById(offerID);
 
         Order order = new Order();
-        ResponseEntity.ok(orderService.createOrder(order, offerID));
+        ResponseEntity.ok(orderService.createOrder(order, offerID, user.getShopperID()));
 
 
 
@@ -61,17 +68,59 @@ public class OrderController {
         return ResponseEntity.ok(orderService.getAllOrders());
     }
 
+    // updating order methods
+
     @PutMapping("/{orderID}")
     public ResponseEntity<Order> updateOrder(@PathVariable Long orderID, @Valid @RequestBody Order order){
         return ResponseEntity.ok(orderService.updateOrder(orderID, order));
     }
 
-    @GetMapping("/delete/{orderID}")
-    public String deleteOrder(@PathVariable Long orderID) {
-        orderService.deleteOrder(orderID);
+    /** endpoint to accept an order
+     * comes from a js fetch on seller page
+     */
+    @PutMapping("/{orderID}/accept")
+    public ResponseEntity<Order> acceptOrder(@PathVariable Long orderID){
+        return ResponseEntity.ok(orderService.changeStatus(orderID, 2));
+    }
+    /** endpoint to reject an order
+     * comes from a js fetch on seller page
+     */
+    @PutMapping("/{orderID}/reject")
+    public ResponseEntity<Order> rejectOrder(@PathVariable Long orderID){
+        return ResponseEntity.ok(orderService.changeStatus(orderID, 3));
+    }
+    /** endpoint to reject an order
+     * comes from a js fetch on shopper page
+     */
+    @PutMapping("/{orderID}/complete")
+    public ResponseEntity<Order> completeOrder(@PathVariable Long orderID){
+        return ResponseEntity.ok(orderService.changeStatus(orderID, 4));
+    }
 
+    /** endpoint to delete an order 
+    * comes from js fetch on shopper page
+    * @param orderID order to delete
+    * @return
+    */
+    @DeleteMapping("/{orderID}")
+    public String deleteOrder(@PathVariable Long orderID, Authentication auth) {
+        // if user is seller
+        if (auth == null || !auth.isAuthenticated()){
+            return "redirect:/login";
+        }
+        if (auth.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_SELLER"))){
+            return "redirect:/403";
+
+        } else if (auth.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_SHOPPER"))){
+            orderService.deleteOrder(orderID);
+            return "redirect:/shoppers/cart";
+        }
+        
+        orderService.deleteOrder(orderID);
         return "redirect:/shoppers/cart";
     }
+
+    // back to getting methods
 
     @GetMapping("/{orderID}")
     public ResponseEntity<Order> getOrder(@PathVariable Long orderID){
